Revert "Merge pull request #35 from MinminC/minjeong_notice_220315"
This reverts commit d42f6799dfb43bea0ec8913c60aa6b49a5b811f5, reversing
changes made to 75234ace23725a0b7dfdebe8ac486186100f9257.

+++ src/main/java/com/kh/firstclass/admin/place/model/vo/Place.java
@@ -14,8 +14,8 @@ public class Place {
 	private int placeNo;
 	private String placeName;
 	private String placeAddress;
-	private double placeLon;
-	private double placeLat;
+	private int placeLon;
+	private int placeLat;
 	private String placeDes;
 	private String placeTags;
 	private String area;
